Correct the interpretation of "Mapped" memory
The "Mapped" total is the kernel's NR_FILE_MAPPED, which is the number
of pages in memory which are mapped, therefore it comes from the cache
and not from the total usage.

Fixes a bug where the 'used' display would go negative if the amount of
mapped memory is high.