Do not ignore removable devices
Removable devices are commonplace; their I/O load is not insignificant.
The previous patches ignores loopback (and raid, md, lwm) which can
be considered a bug as I/O is recorded twice, but I don't think that
it tracks removable devices can be considered a bug, so restore this
behaviour.