Use XPM handling by default
These days the library is commonplace, and we have broken away from
trying to support old platforms that nobody has access to and are
not tested.

This means the next release of xosview will require libXpm, but this
library seems to be as standard as the core of X itself.

Let's be careful before introducing any additional use of libXpm
though.