Default IRQ meter to off
Rather conflicting here, the default has always been 'on' but this
only affected BSD where this meter was available.

I propose this patch to default to off; this makes the Linux case
considered the default and it looks like this has been seen and used
by most people. An IRQ rate meter is not a particularly useful metric
on a modern system.