Fixed inverted logic adjustment for the memory meter on linux.
The comment says that "mapped comes from cached" but it was actually
substracting from field[4] (cached) the value of field[3] (mapped).
It has to be the other way round for proper display!

The symptom that led to this bug fix was the following warning message:

  meter MemMeter had a negative value of -12621901824.000000 for field 4

Once the patch is applied, the warning goes away and the display is
properly reflecting the values that can otherwise by seen via "free"
on linux.