Replace the parsing of /proc/meminfo
We have seen multiple problems with the existing code which manifest
themselves as either sporadic or consistent displays of warnings such
as:

  meter MemMeter had a negative value of xxxxxxxxxxxxxxxxx

The previous code was hard to read due to multiple levels of
indirection; firstly it puts content into struct LineInfo, then maps
these to field_[] values. It's possible the fields are used incorrectly,
and we wanted to audit the comment 'mapped comes from cached'.

Improve the code by:

* Removing the levels of indirection through the arrays of fields
  and using named variables instead

* Remove the intermediate calculations which are done using double;
  it's likely the approximation caused by floating point is the cause
  of the sporadic errors when memory is tight and some values overflow.

* Be consistent to the comment 'cached comes from mapped'; if this is
  true then 'mapped' should be excluded from the calculation of fields_[0].

* Use a simple parser using stdio; the code is already using the
  C library and it allows us to report errors on opening the file
  and my experiences in the other parts of the code showed it to be
  less hungry on resources than fstream (this code is used in a tight
  loop)

The result is a significant decrease in lines of code, increase in
readability and hopefully fixes the above bugs (or at least possible
bugs!)