Allow building with musl
According to ioperm(2) man page, the #include is:

       #include <unistd.h> /* for libc5 */
       #include <sys/io.h> /* for glibc */

We can see that sys/io.h defines ioperm() along with other things,
whereas sys/perm.h only defines ioperm() and does not exist with
musl.

Therefore the conclusion is that this patch builds on musl, and should
not affect builds on glibc.

Tested on:

* Alpine Linux 3.15.0 with musl 1.2.2-r7

* Slackware-current with glibc 2.33